tests/wpt/mozilla/tests/input-events/mouse-text-selection-input-text-selection-event.html

<!DOCTYPE html>
<title>Basic selection of text via the mouse in &lt;input type=text&gt; triggers select event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />

<input id="target" style="font: 50px/1 Ahem;" value="123456">

<script>
promise_test(async test => {
    let targetRect = target.getBoundingClientRect();
    let verticalCenterOfLine = targetRect.top + (targetRect.height / 2);

    // Click somewhere else to avoid double-clicking.
    await new test_driver.Actions()
      .pointerMove(500, 500)
      .pointerDown()
      .pointerUp()
      .send();

    // We should only receive a single "mousemove" event if the cursor does not move.
    const eventWatcher = new EventWatcher(test, target, ["mousedown", "mousemove", "mouseup", "select"]);
    const eventPromise = eventWatcher.wait_for(["mousemove", "mousedown", "select", "mousemove", "select", "mouseup"]);

    await new test_driver.Actions()
      .pointerMove(targetRect.left + 1, verticalCenterOfLine)
      .pointerDown()
      .pointerMove(targetRect.left + (50 * 5), verticalCenterOfLine)
      .pointerUp()
      .send();

    await eventPromise;

    assert_equals(target.selectionStart, 0);
    assert_equals(target.selectionEnd, 5);
    assert_equals(target.selectionDirection, "forward");
    target.setSelectionRange(0, 0);
});
</script>
